+++ bms-entity/src/main/java/com/jiuyescm/bms/asyn/entity/BmsAsynCalcuTaskEntity.java
@@ -22,7 +22,7 @@ public class BmsAsynCalcuTaskEntity implements IEntity {
 	// 业务年月 201901
 	private Integer creMonth;
 	// 任务状态 0-等待 10-处理中 20-成功 30-异常 40-丢弃 50-作废
-	private Long taskStatus;
+	private Integer taskStatus;
 	// 费用总数
 	private Integer feesCount;
 	// 本次任务需要计算的费用总数
@@ -143,7 +143,7 @@ public class BmsAsynCalcuTaskEntity implements IEntity {
 	/**
      * 任务状态 0-等待 10-处理中 20-成功 30-异常 40-丢弃 50-作废
      */
-	public Long getTaskStatus() {
+	public Integer getTaskStatus() {
 		return this.taskStatus;
 	}
 
@@ -152,7 +152,7 @@ public class BmsAsynCalcuTaskEntity implements IEntity {
      *
      * @param taskStatus
      */
-	public void setTaskStatus(Long taskStatus) {
+	public void setTaskStatus(Integer taskStatus) {
 		this.taskStatus = taskStatus;
 	}
